Update to compile and run on more modern versions of FreeBSD.
Many thanks to Miles Lott for preliminary work for this update.

The major change is the complete replacement to the configuration code for
the tunnel interface and the routes.  The code now simply calls ifconfig(8)
and route(8) with the appropriate paramters, instead of using the ioctl and
PF_ROUTE APIs.

The remainder of the changes adjust the data types and the struct alignment.

+++ macip.h
@@ -27,7 +27,7 @@
  * SUCH DAMAGE.
  */
 
-extern int  macip_open (char *zone, u_long net, u_long mask, u_long ns, outputfunc_t o);
+extern int  macip_open (char *zone, uint32_t net, uint32_t mask, uint32_t ns, outputfunc_t o);
 extern void macip_close (void);
 extern void macip_idle (void);
